third_party/WebKit/LayoutTests/fast/dom/Document/adopt-node-with-doctype.html

<!DOCTYPE html>
<html>
<head>
<title>This page tests the DOM document.adoptNode method with DocumentType.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" href="../../../resources/testharness.css">
</head>
<body>
<script>
test(function() {
    var doctype = document.doctype;
    assert_equals(doctype.parentNode, document)
    assert_equals(doctype.ownerDocument, document)
    assert_equals(document.adoptNode(doctype), doctype)
    assert_equals(doctype.parentNode, null)
    assert_equals(doctype.ownerDocument, document)
}, "Explicitly adopting a DocumentType should work.")
</script>
</body>
</html>
